docs(release): design system packages (#49)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @nl-design-system/eslint-config@1.0.0

### Major Changes

- 6c54309: Lint all JavaScript and TypeScript files using the
'nl-design-system/all' rules.